+++ release/update_help_txt.sh
@@ -1,8 +1,0 @@
-script_dir="$(dirname "${BASH_SOURCE[0]}")"
-
-#
-# Using the current environment, which has smart_open installed.
-#
-cd "$script_dir/.."
-python -c 'help("smart_open")' > help.txt
-git commit help.txt -m "updated help.txt"