+++ smart_open/ssh.py
@@ -115,7 +115,8 @@ def open_uri(uri, mode, transport_params):
     parsed_uri = parse_uri(uri)
     uri_path = parsed_uri.pop('uri_path')
     parsed_uri.pop('scheme')
-    return open(uri_path, mode, **parsed_uri, **kwargs)
+    final_params = {**parsed_uri, **kwargs}  # transport_params takes precedence over uri
+    return open(uri_path, mode, **final_params)
 
 
 def _connect_ssh(hostname, username, port, password, connect_kwargs):
@@ -180,7 +181,7 @@ def _maybe_fetch_config(host, username=None, password=None, port=None, connect_k
         if port is None:
             try:
                 port = int(cfg["port"])
-            except (IndexError, ValueError):
+            except (KeyError, ValueError):
                 #
                 # Nb. ignore missing/invalid port numbers
                 #
@@ -226,6 +227,7 @@ def open(
     port=None,
     connect_kwargs=None,
     prefetch_kwargs=None,
+    buffer_size=-1,
 ):
     """Open a file on a remote machine over SSH.
 
@@ -238,7 +240,7 @@ def open(
     mode: str, optional
         The mode to use for opening the file.
     host: str, optional
-        The hostname of the remote machine.  May not be None.
+        The hostname of the remote machine. May not be None.
     user: str, optional
         The username to use to login to the remote machine.
         If None, defaults to the name of the current user.
@@ -251,6 +253,8 @@ def open(
     prefetch_kwargs: dict, optional
         Any additional settings to be passed to paramiko.SFTPFile.prefetch.
         The presence of this dict (even if empty) triggers prefetching.
+    buffer_size: int, optional
+        Passed to the bufsize argument of paramiko.SFTPClient.open.
 
     Returns
     -------
@@ -297,7 +301,7 @@ def open(
             #
             del _SSH[key]
 
-    fobj = sftp_client.open(path, mode)
+    fobj = sftp_client.open(path, mode=mode, bufsize=buffer_size)
     fobj.name = path
     if prefetch_kwargs is not None:
         fobj.prefetch(**prefetch_kwargs)